+++ HISTORY.md
@@ -5,6 +5,7 @@ https://github.com/josdejong/ducktype
 
 ## version 0.1.0
 
+- Added support for structured arrays, like `ducktype([Number])`.
 - Support for combining types, like `ducktype(Number, String)`.
 - Added basic types Array, Object, RegExp, null, undefined.